mod for an edit page

+++ part_pages/part_referee_selector.php
@@ -4,8 +4,17 @@
     $refereeAPIdata = file_get_contents($refereeAPIpath, false, $context);
     $refereeList = json_decode($refereeAPIdata, true);
     
-    echo "<option value='default'>Select Referee</option>";
+    // if this isnt an edit page, make the default to be "select ref"
+    if (!isset($existingRefereeNameToEdit)) {
+        echo "<option value='default'>Select Referee</option>";
+    }
+
+    // modify the output based on the page type (edit or add new)
     foreach($refereeList as $singleRef) {
-        echo "<option value='{$singleRef['refereename']}'>{$singleRef['refereename']}</option>";
+        if (isset($existingRefereeNameToEdit) && $existingRefereeNameToEdit == $singleRef['refereename']) {
+            echo "<option value='{$singleRef['refereename']}' selected='selected'>{$singleRef['refereename']}</option>";
+        } else {
+            echo "<option value='{$singleRef['refereename']}'>{$singleRef['refereename']}</option>";
+        }
     }
 ?>
